Add the skeleton of the HBase client.
This is starting to sketch the outlines of the client, with some initial
basic code to deal with regions, part of the code for region lookups, and
the embryo of a Get RPC.  Much of the code is borrowed from AsyncHBase.

The `cli' tool executes a hard-coded Get RPC and prints the result.

There is no error checking or anything fancy yet.