Dedicated writer thread, deadlines now supported
Each region.Client now has a dedicated goroutine for talking to its
accompanying HBase node. When RPC calls are to be sent, they are added
to a queue that this thread periodically checks for messages to be sent.

Once a message is sent, the results and errors are written back on
channels stored in the RPC struct, that are then returned to the user
that initiated the call.

Deadlines can be set on RPCs now, and the function that waits on the
channels for results from the writer thread will give up and return an
error if the deadline is exceeded. If the deadline is not set, it is
ignored.

Change-Id: I7d2d53e57a47c5f0d06f642b18817ae7c73f3b77