Deadlines handled with context package, single result channel
Now all of the user-facing .*WithDeadline functions generate an
appropriate context, and pass it to c.SendRPC. It's currently only used
to track deadlines, but more data can be added to the context later.

The multiple channels in hrpc.Call have now been combined to a single
channel. A struct type was created to pass multiple values through it at
once.

Change-Id: I0865d3292484b4afcf893c67f3a572527d462ba6