Added some initial tests for Get and Put
integration_test.go has been added to include some basic integration
tests.

The test package has been created to provide functions for tests to
create and delete tables. This is accomplished via the hbase shell, and
as such the HBASE_HOME environment variable must be set to be able to
find the shell.

Change-Id: Ib5b828d97cb69acd2502ba7e60486e48f6b98e5b