Handle network errors, better enforce deadlines.
When an error related to the network happens, be it the server is
unreachable or it sent us a corrupted message, this change has gohbase
flush all queued and sent RPCs with an error and close out the reading
and writing goroutines. client.go, upon seeing a network error, will
kick off a goroutine to repeatedly attempt to find/connect to the server
for the desired region. Once a valid connection is established, all RPCs
waiting for a good connection are woken up to be resent.

More logic has also been added to the creation of region clients in
client.go to enforce the deadline for a request in more cases.

Change-Id: I3daef07048a9de0f5d594811898ac5c0222a7a5e