TestMultiplePutsGets demonstrated a race condition
The RPCs were being written to the wire before their Call IDs were being
recorded by the writer thread in the sentRPCs map, and there was a small
chance of the reader thread receiving HBase's reply before the writer
thread did this. This resulted in the reader thread tossing the message
and logging an error, because it didn't know who to send the reply to.

Change-Id: I540275bcee8506fd74ed27310fe9d8b3fc1eeabe