Fully implemented Filters and Comparators
Effectively created ~30 different types, one for each filter that all
implement a Filter interface. FilterLists are also filters albeit with
additional functionality. FilterLists can be created as such -

1. f := hrpc.NewFilterList(filter1, filter2, filter3)
2. f := hrpc.NewFilterList()
   f.AddFilter(filter1, filter2)
   f.AddFilter(filter3)

Comparators were implemented in the exact same way. All different types
of Comparators implement a Comparator interface.

I've tested a handpicked selection of Filters (FilterList tested
working too), but all follow a similar template so are expected to be
fully working.

Note: The last 7-8 lines of every ConstructPBFilter() method is the same
but I'm unsure of how to pull it out into a separate function because of
the different types.

Change-Id: I29bb7cf93b7e08dccbf7c9b073fe7a8e8e312b82
Signed-off-by: Benoit Sigoure <tsunanet@gmail.com>