Fixes so that TestChangingRegionServers passes.
metaRegionInfo is altered to mark when the metaClient is unavailable,
and thus can no longer be a global variable. This commit moves it to be
a field in the Client struct in client.go, and it is initialized with
the same values in the NewClient constructor.

When reestablishRegion in client.go is called with the metaRegionInfo,
it needs to use locateMeta to reestablish a connection to the meta
region. If it's not handled specially, locateRegion will attempt to use
the meta client to look up its location, and the metaClient cannot be
used to locate the metaClient for obvious reasons.

This commit also removes clients from the region client cache when they
become unavailable.

Change-Id: I3d3738d330bc42eb846c28ab98f914e603b99c26