Fixes so that TestChangingRegionServers passes.
The meta table is used when locateRegion is called, which is done by
reestablishRegion in client.go. If the meta table becomes inaccessible,
reestablishRegion will never succeed for metaRegionInfo because it tries
to use metaRegionInfo to reestablish the connection. Instead of kicking
off a new reestablishRegion for the meta table, it should just fail and
make the client nil, so the next attempt by reestablishRegion will
attempt to refind the meta table.

Change-Id: I3d3738d330bc42eb846c28ab98f914e603b99c26