First pass at aliasing pb response types.
Ran into a lot of difficulty with how responses hold cells. I included
two options for converting the cells to their aliased types. Thoughts
are appreciated.

Also - for Get/Scan requests do we want to return a 'Result' type that
contains 'Cells' as a member or do we want to return the slice of cells
directly?

Change-Id: I2cf94dc015c7b39b57594ee057d3eed15cd1b599