Added admin client, disable and delete table calls.
admin.go has been added, and it contains the admin client. The admin
client exists for sending RPC calls specifically to the HMaster, and
will support functions modifying tables and schemas and stuff like that.
Currently it supports functions for disabling and deleting tables.

The create table function is also provided in this commit, but is
currently non-functional. Upon use, it appears to successfully create
the table in the HMaster, but an error is encountered when the HMaster
attempts to assign the new regions to region servers, rendering the new
table unusable.

Change-Id: I3c29f937d634575187782a915767991d13493080