Adds a reflection-based API for putting new values into HBase.
Adds alternate functions to hrpc/mutate.go for modifying data. Instead
of taking nested maps of byte slices, structs can be passed in. The
exposed fields of the struct will be inspected for hbase tags containing
column families and qualifiers, and these fields are then converted into
byte slices and entered into the protobuf messages as if they had come
from the nested maps like before.

An accompanying commit containing calls for deserializing a pb.Result
into a tagged struct will be written.

The nested maps are still exposed because the way the new function goes
about serializing the fields of the struct may not be the most effective
method, depending on the use case.

Change-Id: I9710be0a483d4c8ade5bc65baa199ac2e80b1707