Refactoring of client.go to simplify reasoning through the code.
client.go was getting incredibly hard for me to follow, so I
consolidated all the logic that a RPC request goes through into a single
function. In my opinion this makes it far easier to see how the logic
between the various steps works, and will make it easier to reason
through race conditions.

I also combined the functionality of admin.go into client.go. There was
too much duplication of logic between the two files to make the
separation worth it.

Having a separate function for each possible RPC in client.go felt
absurd, because all of them were doing the exact same thing, so I
removed them. The resulting interface is slightly less convenient due to
the casting of results required, but this will be easier to maintain.

Change-Id: I5f7e42b359ad62e33d98fb2d46383eef7b57e867