Update patch set 1
Abandoned

Oops, I didn't mean to push this one.  It was still WIP, and I found a weird problem with HBase: when I disable a table it marks the regions as disabled in ZooKeeper but not in meta.  So on our end we don't see the regions of the table as offline, we merely keep getting NSRE's when trying to access them.

Patch-set: 1
Status: abandoned