Eliminate the goto, and fix a bug where we read from a nil chan
In sendRPCToRegion, it drops into that big if block if ch == nil. None
of the goto's that jump outside of the if block set ch to be anything
other than nil. When the select statement reads from nil, it goes to
sleep forever (or until the deadline).

Why the behavior of reading from a nil channel is to sleep forever
baffles me.

Change-Id: I6b2bef7f741a74ddd87ea951d31b3fcc5b513297