Add the go race detection to our tests, resolved issues it found
This commit makes us obey the -race flag on go test. This includes
a performance hit ("memory usage may increase by 5-10x and execution
time by 2-20x"), so it shouldn't be put on the benchmarks.

The issues it found all stem from global variables we were both
reading from and writing to, so I wrapped them in locks. I don't know if
this was strictly necessary, but https://golang.org/ref/mem advised me
to not "be clever", so better safe than sorry.

Change-Id: Id8e4d9c595528b90ea6112e19dcfc892635a21ce