Implements buffered increment operations.
Uses an LRU cache to hold on to increment operations, and combine
together increment operations on identical cells in HBase. The cache is
flushed periodically, to prevent operations from sitting in there
indefinitely.

If a user wishes to bypass the batching logic, increments can still be
sent as before.

Deadlines are not observed on buffered increments.

Change-Id: Ie32ab0ba0c3931a079ac13f176ec0e2e090f42be