Batch together get and mutate RPCs into MultiRequests.
When the region client goes to flush its queue, all get/mutate requests
to the same region are combined together into a single MultiRequest
message. When the MultiResponse is received, the various responses are
then sent back across the channels to the individual get/mutate
requests.

If an error is encountered with a MultiRequest, all get/mutate requests
that were combined receive the error.

Change-Id: I8f79b14d43dcae4ac517af02f28442ea653483b3