Send a probe to region before marking it as available
When establishRegion successfully acquires a client for a region (it
could come from the cache or make a new one), it will send a exists get
request for a key in the region. The region is then only marked as
available if this request returns without error.

Also checks if region must be added to cache when a valid client is
found in the cache.

Change-Id: I7ab449e7137d01778e934871eca257efc921113b