gohbase, gohbase/hrpc: request partial results for scan
Since rows in HBase can have "unlimited" number of columns,
it's not a good idea to fetch an entire row at a time. We can
use client_handles_partial_results to ask for parts of a row
in order not to put too much presure on regionserver and
network. The row is still reconstructed on the client side and
there will be an option to ask for partials later.

Change-Id: I7e1bd79292cfceeb0960a612dee90cd9707da665