gohbase, gohbase/hrpc: improve scanning interface
The old Scan was not correct as it could only read NumberOfRows a time
and would return all the results at once. The caller had to recreate the
scan every time NumberOfRows is reached. The new implementation allows
to scan arbitrary large tables without the need to recreate the Scan.

Another cool side-effect of using golang is that the implementation
turned out to be thread-safe. Which means that as long as results don't
need to be ordered, the cached results can been processed in parallel.

Change-Id: I98389fffc4ae9891bedfaee09c9c4564a280a1a2