gohbase: don't close the RegionClient if it has no regions
Just because we think there's no more regions at regionserver, doesn't
mean that we should discard the connection to it. Maybe regionserver had
only one region that was split, so we first removed the old one and add
a new one that would result in unnecessary client reestablishment. Or maybe
reginserver was temporarily drained. If the regionserver is alive, we
should keep the connection open.

Change-Id: I94a7523d6fd5ff497a6dad947e7d8d68b37ad4cc