gohbase: remove client from cache if probe fails with UnrecoverableError
There's a bug that if a regionserver is stuck and hbase:meta hasn't been
updated yet to exclude it, we would add it to the cache and the first
probe would fail with UnrecoverableError. However, since we didn't
remove the client from cache, the probe would keep grabbing it over and
over without any success.

Change-Id: Ice552b9d427ecf1cd1da2fe9b5c8703dc1d99b22