gohbase: fix meta lookup rate limiting and improve test
If a request has been throttled, it's not always appropriate
to look into region cache again in case some other goroutine
added our region. In case we marked region in cache as unavailble,
if meta lookup rpc is throttled, we will get the same region we are
trying to reestablish for cache.

Also, configured the Limiter to have burst size of 100 to allow
a hundred concurrent requests at any moment.

Change-Id: I368976b85750f419e4a958730d4f0c4bca47b8be