test: Fix TestCreateTable
This test checks if one region is created for the
newly created table. The prefix scan for this should
be 'testTableName + ","' and not 'testTableName + ",,"'
(note the second comma).
Using two commas will always return one row even if the
table has multiple regions as only one region is the
start region and having two commas will return only
the start region.
So, the prefix is changed to scan with only one comma

Change-Id: I55fa53f3e685470aa247d7112090473e170a8195