gohbase/region: return rs unrecoverable exception to the caller
This is a sloppy bug that would result in stuck rpcs to be if an
unrecoverable exception is returned by regionserver. The rpc would
already be unregistered, so a call to fail() would not dispatch
error to it. Using defer will make sure that all errors are returned
to the caller.

Change-Id: I36f77a5b8ba33501235e612746ea42f906ed805a