hrpc, gohbase: Add snapshot creation and deletion
This commit introduces the ability to add and delete snapshots via the
admin client.

Due to the snapshot API being slightly different to other admin calls,
we introduce a new option SnapshotValidateInterval which allows us to
know when the Snapshot has been created.

More features will be introduced in the future to ensure snapshots that
are created are valid.