region: reestablish region on java.io.IOException: Cannot append; log is closed
Regionserver 2.2.x can throw this exception when trying to do mutate on a
region where log has been closed due to not being able to not being able to
recover lease in hdfs. Ideally regionserver should have never returned this error
to clients, but since it does we need to deal with it instead of propagating this
error all the way to users of gohbase.

Change-Id: I8bb7a06d965424d2e24812cc4f15cddfe335b42c