hrpc: expose CacheBlocks option for scan/get queries
This option allows the user to specify if they want
blocks read by the query to be pulled into the cache,
which in some cases (e.g. full table scan), may be
useful.

The default is true as defined in the models for Get and
Scan in pb/Client.proto