region: fix TestProcessRPCs/2 test
The test was expecting that the scheduler would let
processRPCs goroutine to read several RPCs before moving on
to next goroutine. In case the sheduler lets processRPC to only
process one RPC at a time, we would end up with a call to regionserver
for each RPC (no batching). This can be especially reproduced with
-race flag.

Instead we should expect that at most there would be as many batches
as RPCs.