integration: fix TestMaxResultsPerColumnFamilyScan test result
The expected number of rows was wrongly set to 1 instead of 2. The test
passed until today because the version of HBase used by the CI (HBase
2.2.3) was affected by a bug when the filter MaxResultPerColumnFamily is
used in a scan request.

More information about the bug, the fix and affected/fixed version in
the HBase JIRA ticket: https://issues.apache.org/jira/browse/HBASE-26289