region/client: Move cancel check from trySend to QueueRPC
trySend is called from both QueueRPC and processRPCs. processRPCs is
already selecting on the done context for the RPC, so no need to do it
again. In addition the context on a multi is never cancelled, so it's
not useful to check.

Arguably we don't need this check at all here since there is no
blocking operation in this case of QueueRPC. The best reason to keep
it is to give a slightly better error message if you try to write to a
shut down region server.