region: Remove independent span created in processRPCs
It is confusing to have a different span here and not likely to be
used by viewers of the data. To track the time an RPC spends in the
processRPCs loop we can add a span to that RPC's context as it enters
the loop and close the span when the RPC is flushed.

Once the span is removed from processRPCs, much of the code related to
open telemetry tracing is no longer needed in the region package.