region: Add batch_flush_size histogram metric
This will expose the size of batches being sent by gohbase.

Buckets are [1, 5, 25, 125, 625, 3125, 15625, 78125]. This should
capture all expected sizes of batches and importantly count all the
batches that have exactly 1 request in them. In one setup I saw 90% of
batches containing exactly 1 request and another the 99.9%tile value
was ~3100 with spikes over 8000.