region: Reimplement processRPCs batching loop
The old implementation had lots of different conditions that were hard
to think through, some uninuitive behaviors, and at least one bug.

First the bug: when flushInterval == 0, the flush reason is always
"queue full". It should be "timeout" if we flush before filling the
batch. This is fixed in the new loop.

Old implementation if flushInterval == 0:
1) do a nonblocking read of rpcs chan in a loop
2) check if you read anything in that loop
2a) if not, do a blocking read to get at least one RPC
    and then go back to 1.
2b) if yes, flush and start over

Old implementation if flushInterval > 0:
1) do a nonblocking read of rpcs chan in a loop
2) check if you read anything in that loop
2a) if not, do a blocking read to get at least one RPC
    and then go back to 1.
2b) if yes, and size of batch is >= than rpcQueueSize
    then flush and start over
2c) if yes and size of batch is < rpcQueueSize
    then start a timer and read rpcs chan
    until timer finishes or batch is full
    then flush and start over

The unintuitive parts are the loop back to step 1, and when
flushInterval > 0 the timer is potentially started well after it has
started adding RPCs to the batch.

The new implemenation:

1) Do a blocking read of rpcs chan once, so we get at least one RPC in
   our batch.
2) If the batch has exceeded the rpcQueueSize, flush and start over.
3a) If flushInterval == 0
    Start loop that does non-blocking read of rpcs chan
    Once read is unsuccessful then flush and start over
3b) If flushInterval > 0
    Start a timer and read rpcs chan
    until timer finishes or batch is full
    then flush and start over