rpc: refactor for adding batch support
* Factor out response error handling into a function

This code will be reused by the Batch code.

* Add getRegionAndClientForRPC

getRegionAndClientForRPC gets the region for an RPC and then gets
client. If either is not available it will launch a goroutine to make
the region available and set its client and will wait for it to be
available or a context is cancelled.

Finding the region's client and the error handling was split between
SendRPC and sendRPCToRegion. This refactors it into one function that
can be used by both SendRPC and SendBatch.