SendBatch: Don't start goroutines
Instead of starting goroutines we can queue each region server's
collection of RPCs and only after queueing them all wait for all
responses. We get a similar amount of parallelism, the only potential
slow down is that we can't queue one region server's RPCs until the
previous is able to be queued.

By not starting goroutines, the resource usage of SendBatch is more
predictable. And not starting goroutines is likely to be less
expensive overall.