region: Simplify buffer funcs
Replace `growBuffer` with 'append' and `resizeBufferCap` with
`slices.Grow`. And remove redundant 'if' checks on capacity or size
that `append` will do as well. This is equivalent behavior after
undoing the unintentional change from commit 1fee39f343954. Benchmarks
show the performance is equivalent to the code before commit
1fee39f343954.

Also replace some deprecated calls to rand funcs.