rpc: ignore offline/unassigned regions during cache warmup
When warming up the cache, ignore regions that are offline/unassigned.
They could exist for many reasons:
- parent region after a split, until daughter regions are compacted
- regions that are merged together, until compaction happens
- region moving due to HBase LoadBalancer, RegionServer shutting down or
  manually
- etc

It's not critical if we are missing a few regions in the cache during
warmup. The most important thing is for the warmup to be fast and to
fill in with as many regions as possible. Other regions will be fetched
later as needed.