Don't restart scanner renewer on every Next call
Most calls to Next are just reading some buffered results. The scanner
renewer only needs to be stopped and started again when we ask for
more results from HBase. Also, we shouldn't start a renewer if we know
we have already fetched the final results.

Pass in client's logger to scanner.

Use renewLoop's ctx in renew().

Doubled scannerLease sleeps in integration tests to reduce flakiness
of tests.