region: Set default KeepAliveConfig on regionserver connections
Configures default KeepAliveConfig on regionserver connections,
including setting TCP_USER_TIMEOUT on linux systems.

TCP_USER_TIMEOUT is an additional layer of protection. It handles the
case of a dead connection, but the send buffer is full such that keep
alives packets are not able to be sent. By setting it to the time of
the full timeout of the normal keepalives it won't short-circuit the
normal keepalives.

See https://blog.cloudflare.com/when-tcp-sockets-refuse-to-die/