region: Set TCP_USER_TIMEOUT on regionserver connections
This is an check for a dead connection that handles the case the send
buffer is full, so that keep alives packets are not sent. By setting
it to the time of the full timeout of the normal keepalives it won't
short-circuit the normal keepalives while adding another layer of
protection. See
https://blog.cloudflare.com/when-tcp-sockets-refuse-to-die/