rpc: clarify logging for ScanStats with ScanMetrics
When a scan attempt returns an error that originates from sendBlocking,
there won't be a ScanResponse (the RPCResult will be nil).
In this case when updating the ScanStats and while ScanMetrics are
enabled, an unclear error level log message was logged.
This change clarifies the log message and moves the log level to debug,
skips the check when there's a nil RPCResult, and adds more testing.