hrpc: Add Region and MoreResultsInRegion fields to ScanResponseV2
A Scan that returns no responses doesn't provide any information about
how far the Scan processed. This may be useful if doing a broad Scan.
One way to know this information is if a response has no more results
in the region, then we know that the Scan has progressed at least to
the EndKey of that region.